+++ server/src/routes/workAreas.routes.ts
@@ -1,6 +1,7 @@
 import { Router } from 'express'
 import { getDb } from '../db/database.js'
 import { jsonErr, jsonOk } from '../utils/http.js'
+import { requirePermission } from '../middleware/stationAuth.js'
 import * as workAreaService from '../services/workAreaService.js'
 
 export const workAreasRouter = Router()
@@ -8,7 +9,8 @@ export const workAreasRouter = Router()
 workAreasRouter.get('/', (req, res) => {
   try {
     const stationId = typeof req.query.stationId === 'string' ? req.query.stationId : undefined
-    jsonOk(res, workAreaService.listWorkAreas(getDb(), stationId))
+    if (!requirePermission(req, res, stationId, 'settings.view')) return
+    jsonOk(res, workAreaService.listWorkAreas(getDb(), stationId!))
   } catch (e) {
     jsonErr(res, e instanceof Error ? e.message : 'Fehler', 500)
   }
@@ -16,6 +18,9 @@ workAreasRouter.get('/', (req, res) => {
 
 workAreasRouter.get('/:id', (req, res) => {
   try {
+    const sid = workAreaService.getWorkAreaStationId(getDb(), req.params.id)
+    if (!sid) return jsonErr(res, 'Arbeitsbereich nicht gefunden', 404)
+    if (!requirePermission(req, res, sid, 'settings.view')) return
     const w = workAreaService.getWorkArea(getDb(), req.params.id)
     if (!w) return jsonErr(res, 'Arbeitsbereich nicht gefunden', 404)
     jsonOk(res, w)
@@ -27,7 +32,8 @@ workAreasRouter.get('/:id', (req, res) => {
 workAreasRouter.post('/', (req, res) => {
   try {
     const stationId = typeof req.query.stationId === 'string' ? req.query.stationId : undefined
-    jsonOk(res, workAreaService.createWorkArea(getDb(), req.body ?? {}, stationId), 201)
+    if (!requirePermission(req, res, stationId, 'settings.edit')) return
+    jsonOk(res, workAreaService.createWorkArea(getDb(), req.body ?? {}, stationId!), 201)
   } catch (e) {
     jsonErr(res, e instanceof Error ? e.message : 'Fehler', 400)
   }
@@ -35,6 +41,9 @@ workAreasRouter.post('/', (req, res) => {
 
 workAreasRouter.put('/:id', (req, res) => {
   try {
+    const sid = workAreaService.getWorkAreaStationId(getDb(), req.params.id)
+    if (!sid) return jsonErr(res, 'Arbeitsbereich nicht gefunden', 404)
+    if (!requirePermission(req, res, sid, 'settings.edit')) return
     jsonOk(res, workAreaService.updateWorkArea(getDb(), req.params.id, req.body ?? {}))
   } catch (e) {
     jsonErr(res, e instanceof Error ? e.message : 'Fehler', 400)
@@ -43,6 +52,9 @@ workAreasRouter.put('/:id', (req, res) => {
 
 workAreasRouter.delete('/:id', (req, res) => {
   try {
+    const sid = workAreaService.getWorkAreaStationId(getDb(), req.params.id)
+    if (!sid) return jsonErr(res, 'Arbeitsbereich nicht gefunden', 404)
+    if (!requirePermission(req, res, sid, 'settings.edit')) return
     workAreaService.deleteWorkArea(getDb(), req.params.id)
     jsonOk(res, { deleted: true })
   } catch (e) {
